refactor(training,#14363): extract bias/MSE helpers to shared torch-free module
Move _mse_decomposition and _dm_centered_mse from btc_vol.py (canonical),
btc_m15.py and hmm_regime_vol.py into scripts/bias_metrics.py -- a module
with no torch dependency so CPU-only scripts can import it (hmm_regime_vol
is HMM+OLS only; importing btc_vol would drag torch via dlinear_vol).
4th consumer btc_patchtst.py re-pointed. Outputs proven byte-identical on
a 9-entry reference battery (NaN-aware comparison); 90/90 tests pass.
Also add encoding="utf-8" to btc_m15 run_smoke subprocess call (pre-commit
check-subprocess-encoding, #12811).

Closes #14363
